requirements.txt
google-genai
pillow
moviepy
python-dotenv
flask
openai
together
numpy